requirements.txt
requests
python-dotenv
convertapi
numpy
